bind FileReader with FileResource
