+++ lazyllm/tools/rag/dataReader.py
@@ -269,11 +269,24 @@ config.add('rag_filename_as_id', bool, False, 'RAG_FILENAME_AS_ID')
 
 
 class FileReader(object):
+    def __init__(self, file_resource: Optional[Callable] = None) -> None:
+        self.file_resource = file_resource
 
     def __call__(self, input_files):
-        file_list = _lazyllm_get_file_list(input_files)
-        if isinstance(file_list, str) and file_list is not None:
+        if input_files:
+            file_list = _lazyllm_get_file_list(input_files)
+        else:
+            file_list = []
+        if isinstance(file_list, str):
             file_list = [file_list]
+        if self.file_resource is not None:
+            extra_file = self.file_resource()
+            if isinstance(extra_file, str):
+                file_list.append(extra_file)
+            elif isinstance(extra_file, list) and all(isinstance(item, str) for item in extra_file):
+                file_list.extend(extra_file)
+            else:
+                LOG.warning(f'not support type:{type(extra_file)}')
         if len(file_list) == 0:
             return []
         nodes = SimpleDirectoryReader(input_files=file_list)._load_data()
