+++ tests/advanced_tests/standard_test/test_engine.py
@@ -250,3 +250,14 @@ class TestEngine(object):
 
         r = engine.run(gid, "这张图片描述的是什么？")
         assert '.wav' in r
+
+    def test_OCR(self):
+        nodes = [dict(id='1', kind='OCR', name='m1', args=dict(model="PP-OCRv5_mobile"))]
+        edges = [dict(iid='__start__', oid='1'), dict(iid='1', oid='__end__')]
+        data_root_dir = os.getenv("LAZYLLM_DATA_PATH")
+        input = os.path.join(data_root_dir, "rag_master/default/__data/pdfs/reading_report_p1.pdf")
+        engine = LightEngine()
+        gid = engine.start(nodes, edges)
+        data = engine.run(gid, input)
+        verify = lazyllm.components.ocr.pp_ocr.OCR("PP-OCRv5_mobile")(input)
+        assert len(data) == len(verify)
